+++ table.h
@@ -95,7 +95,7 @@ size_t tableCapacity(Table *table);
   }
 #endif
 
-struct ObjString *tableFindString(Table* table, const char* chars, int length,
+struct ObjString *tableFindString(Table* table, const char* chars, size_t length,
                            uint32_t hash);
 
 void tableRemoveWhite(Table *table);
